fix: restructure app
