--- components/static/GameHeader.tsx
@@ -1,7 +1,0 @@
-export default function GameHeader() {
-  return (
-    <header className="p-12 flex flex-col space-y-2 text-center">
-      <h1 className="scroll-m-20 text-4xl font-bold tracking-tight lg:text-5xl lg:text-extrabold mx-auto">TIC TAC TOE GAME</h1>
-    </header>
-  );
-}
